[refactor/#339] 추천 장소 컴포넌트를 RecommendedPreviewContentList로 분리
